refactor: parseGitAuthor to util/git
