+++ lib/platform/index.ts
@@ -1,10 +1,10 @@
 import URL from 'url';
-import addrs from 'email-addresses';
 import type { AllConfig } from '../config/types';
 import { PLATFORM_NOT_FOUND } from '../constants/error-messages';
 import { logger } from '../logger';
 import type { HostRule } from '../types';
 import { setNoVerify, setPrivateKey } from '../util/git';
+import { parseGitAuthor } from '../util/git/author';
 import * as hostRules from '../util/host-rules';
 import platforms from './api';
 import type { Platform } from './types';
@@ -37,48 +37,6 @@ export function setPlatformApi(name: string): void {
     );
   }
   _platform = platforms.get(name);
-}
-
-interface GitAuthor {
-  name?: string;
-  address?: string;
-}
-
-export function parseGitAuthor(input: string): GitAuthor | null {
-  let result: GitAuthor = null;
-  if (!input) {
-    return null;
-  }
-  try {
-    result = addrs.parseOneAddress(input);
-    if (result) {
-      return result;
-    }
-    if (input.includes('[bot]@')) {
-      // invalid github app/bot addresses
-      const parsed = addrs.parseOneAddress(
-        input.replace('[bot]@', '@')
-      ) as addrs.ParsedMailbox;
-      if (parsed?.address) {
-        result = {
-          name: parsed.name || input.replace(/@.*/, ''),
-          address: parsed.address.replace('@', '[bot]@'),
-        };
-        return result;
-      }
-    }
-    if (input.includes('<') && input.includes('>')) {
-      // try wrapping the name part in quotations
-      result = addrs.parseOneAddress('"' + input.replace(/(\s?<)/, '"$1'));
-      if (result) {
-        return result;
-      }
-    }
-  } catch (err) /* istanbul ignore next */ {
-    logger.error({ err }, 'Unknown error parsing gitAuthor');
-  }
-  // give up
-  return null;
 }
 
 export async function initPlatform(config: AllConfig): Promise<AllConfig> {
